- Admin reject/accept sellers - User update id to Long - User role Fetch mode Eager - Add Security constants for easy access

+++ src/main/java/com/project/onlinestore/utils/SecurityConstants.java
@@ -1,7 +1,7 @@
 package com.project.onlinestore.utils;
 
 public class SecurityConstants {
-    final public static int ROLE_ADMIN = 1;
-    final public static int ROLE_BUYER = 3;
-    final public static int ROLE_SELLER = 2;
+    final public static long ROLE_ADMIN = 1;
+    final public static long ROLE_BUYER = 3;
+    final public static long ROLE_SELLER = 2;
 }
